app/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
requests==2.31.0
jinja2==3.1.2